Revert some changes to joins
